refactor: todoapp directory
