+++ todoapp/repository/todo_repository_test.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/podhmo-sandbox/sample-api/entity"
+	"github.com/podhmo-sandbox/sample-api/todoapp/entity"
 	"github.com/podhmo-sandbox/sample-api/pkg/dblib"
-	"github.com/podhmo-sandbox/sample-api/repository/repositorytest"
+	"github.com/podhmo-sandbox/sample-api/todoapp/repository/repositorytest"
 	_ "modernc.org/sqlite"
 )
